requirements.txt
streamlit
groq
python-dotenv
tenacity
requests
huggingface_hub
validators
pandas
yara-python
python-whois
